requirements.txt
watchdog==3.0.0
web3==6.11.1
cryptography==41.0.4
python-magic==0.4.27
psutil==5.9.5
requests==2.31.0
python-dotenv==1.0.0
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
joblib==1.3.2